Combined the classes into one file, started making methods for piece move validity
